refactor: hide backend module
